# docs/swagger.yaml
basePath: /api
definitions:
...
  domain.Apartment:
    properties:
      agreement_accepted_at:
        type: string
      amenities:
        items:
          $ref: '#/definitions/domain.PopularAmenities'
        type: array
      apartment_number:
        type: integer
      apartment_type:
        $ref: '#/definitions/domain.ApartmentType'
      apartment_type_id:
        type: integer
      booking_count:
        type: integer
      building:
        type: string
      city:
        $ref: '#/definitions/domain.City'
      city_id:
        type: integer
      condition:
        $ref: '#/definitions/domain.ApartmentCondition'
      condition_id:
        type: integer
      contract_id:
        type: integer
      created_at:
        type: string
      daily_price:
        type: integer
      description:
        type: string
      district:
        $ref: '#/definitions/domain.District'
      district_id:
        type: integer
      documents:
        items:
          $ref: '#/definitions/domain.ApartmentDocument'
        type: array
      floor:
        type: integer
      house_rules:
        items:
          $ref: '#/definitions/domain.HouseRules'
        type: array
      id:
        type: integer
      is_agreement_accepted:
        type: boolean
      is_favorite:
        type: boolean
      is_free:
        type: boolean
      kitchen_area:
        type: number
      listing_type:
        type: string
      location:
        $ref: '#/definitions/domain.ApartmentLocation'
      microdistrict:
        $ref: '#/definitions/domain.Microdistrict'
      microdistrict_id:
        type: integer
      moderator_comment:
        type: string
      owner:
        $ref: '#/definitions/domain.PropertyOwner'
      owner_id:
        type: integer
      photos:
        items:
          $ref: '#/definitions/domain.ApartmentPhoto'
        type: array
      price:
        type: integer
      rental_type_daily:
        type: boolean
      rental_type_hourly:
        type: boolean
      residential_complex:
        type: string
      room_count:
        type: integer
      service_fee_percentage:
        type: integer
      status:
        $ref: '#/definitions/domain.ApartmentStatus'
      street:
        type: string
      total_area:
        type: number
      total_floors:
        type: integer
      updated_at:
        type: string
      view_count:
        type: integer
    type: object
...
  domain.Booking:
    properties:
      apartment:
        $ref: '#/definitions/domain.Apartment'
      apartment_id:
        type: integer
      booking_number:
        type: string
      can_extend:
        type: boolean
      cancellation_reason:
        type: string
      cleaning_duration:
        type: integer
      contract_id:
        type: integer
      created_at:
        type: string
      door_status:
        $ref: '#/definitions/domain.DoorStatus'
      duration:
        type: integer
      end_date:
        type: string
      extension_duration:
        type: integer
      extension_end_date:
        type: string
      extension_price:
        type: integer
      extension_requested:
        type: boolean
      final_price:
        type: integer
      id:
        type: integer
      is_contract_accepted:
        type: boolean
      last_door_action:
        type: string
      owner_comment:
        type: string
      payment_id:
        type: integer
      renter:
        $ref: '#/definitions/domain.Renter'
      renter_id:
        type: integer
      service_fee:
        type: integer
      start_date:
        type: string
      status:
        $ref: '#/definitions/domain.BookingStatus'
      total_price:
        type: integer
      updated_at:
        type: string
    type: object
...
  domain.ChatRoom:
    properties:
      apartment:
        $ref: '#/definitions/domain.Apartment'
      apartment_id:
        type: integer
      booking:
        $ref: '#/definitions/domain.Booking'
      booking_id:
        type: integer
      closed_at:
        type: string
      concierge:
        $ref: '#/definitions/domain.Concierge'
      concierge_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      last_message_at:
        type: string
      opened_at:
        type: string
      renter:
        $ref: '#/definitions/domain.Renter'
      renter_id:
        type: integer
      status:
        $ref: '#/definitions/domain.ChatRoomStatus'
      updated_at:
        type: string
    type: object
  domain.ChatRoomStatus:
    enum:
    - pending
    - active
    - closed
    - archived
    type: string
    x-enum-varnames:
    - ChatRoomStatusPending
    - ChatRoomStatusActive
    - ChatRoomStatusClosed
    - ChatRoomStatusArchived
...
  domain.Concierge:
    properties:
      apartments:
        items:
          $ref: '#/definitions/domain.Apartment'
        type: array
      created_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      schedule:
        $ref: '#/definitions/domain.ConciergeSchedule'
      updated_at:
        type: string
      user:
        $ref: '#/definitions/domain.User'
      user_id:
        type: integer
    type: object
...
  domain.CreateBookingRequest:
    properties:
      apartment_id:
        type: integer
      duration:
        minimum: 1
        type: integer
      start_date:
        type: string
    required:
    - apartment_id
    - duration
    - start_date
    type: object
  domain.CreateChatRoomRequest:
    properties:
      booking_id:
        type: integer
    required:
    - booking_id
    type: object
...
  domain.CreateConciergeRequest:
    properties:
      apartment_ids:
        items:
          type: integer
        minItems: 1
        type: array
      schedule:
        $ref: '#/definitions/domain.ConciergeSchedule'
      user_id:
        type: integer
    required:
    - apartment_ids
    - user_id
    type: object
...
  domain.ErrorResponse:
    properties:
      error:
        type: string
      success:
        type: boolean
    type: object
...
  domain.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  domain.UpdateConciergeRequest:
    properties:
      is_active:
        type: boolean
      schedule:
        $ref: '#/definitions/domain.ConciergeSchedule'
    type: object
...
  http.CreateApartmentRequest:
    properties:
      amenity_ids:
        items:
          type: integer
        type: array
      apartment_number:
        type: integer
      building:
        type: string
      city_id:
        type: integer
      condition_id:
        type: integer
      daily_price:
        minimum: 0
        type: integer
      description:
        type: string
      district_id:
        type: integer
      floor:
        minimum: 1
        type: integer
      house_rule_ids:
        items:
          type: integer
        type: array
      kitchen_area:
        type: number
      latitude:
        type: number
      listing_type:
        enum:
        - owner
        - realtor
        type: string
      longitude:
        type: number
      microdistrict_id:
        type: integer
      photos_base64:
        items:
          type: string
        type: array
      price:
        minimum: 0
        type: integer
      rental_type_daily:
        type: boolean
      rental_type_hourly:
        type: boolean
      residential_complex:
        type: string
      room_count:
        minimum: 1
        type: integer
      street:
        type: string
      total_area:
        type: number
      total_floors:
        minimum: 1
        type: integer
    required:
    - apartment_number
    - building
    - city_id
    - condition_id
    - district_id
    - floor
    - kitchen_area
    - listing_type
    - room_count
    - street
    - total_area
    - total_floors
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API для платформы аренды квартир Renti.kz
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Renti.kz API
  version: "1.0"
paths:
...
  /admin/concierges:
    get:
      description: Получает список всех консьержей с фильтрацией и пагинацией (только
        для админов)
      parameters:
      - description: Фильтр по активности
        in: query
        name: is_active
        type: boolean
      - description: Фильтр по ID квартиры
        in: query
        name: apartment_id
        type: integer
      - default: 1
        description: Номер страницы
        in: query
        name: page
        type: integer
      - default: 20
        description: Размер страницы
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список консьержей
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/domain.Concierge'
                type: array
              pagination:
                type: object
              success:
                type: boolean
            type: object
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Список консьержей
      tags:
      - concierge
    post:
      consumes:
      - application/json
      description: Создает нового консьержа и привязывает его к квартире (только для
        админов)
      parameters:
      - description: Данные для создания консьержа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CreateConciergeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Консьерж успешно создан
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.Concierge'
              type: object
        "400":
          description: Неверные данные запроса
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Создание консьержа
      tags:
      - concierge
  /admin/concierges/{id}:
    delete:
      description: Удаляет консьержа (только для админов)
      parameters:
      - description: ID консьержа
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Консьерж удален
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Ошибка при удалении
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Консьерж не найден
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Удаление консьержа
      tags:
      - concierge
    get:
      description: Получает информацию о консьерже по его идентификатору (только для
        админов)
      parameters:
      - description: ID консьержа
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Консьерж найден
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.Concierge'
              type: object
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Консьерж не найден
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получение консьержа по ID
      tags:
      - concierge
    put:
      consumes:
      - application/json
      description: Обновляет информацию о консьерже (только для админов)
      parameters:
      - description: ID консьержа
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateConciergeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Консьерж обновлен
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Неверные данные запроса
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Консьерж не найден
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Обновление консьержа
      tags:
      - concierge
...
  /apartments:
    get:
      consumes:
      - application/json
      description: Получает список квартир с возможностью фильтрации и пагинации
      parameters:
      - description: ID города
        in: query
        name: city_id
        type: integer
      - description: ID района
        in: query
        name: district_id
        type: integer
      - description: ID микрорайона
        in: query
        name: microdistrict_id
        type: integer
      - description: Количество комнат
        in: query
        name: room_count
        type: integer
      - description: Минимальная площадь
        in: query
        name: min_area
        type: number
      - description: Максимальная площадь
        in: query
        name: max_area
        type: number
      - description: Минимальная цена
        in: query
        name: min_price
        type: integer
      - description: Максимальная цена
        in: query
        name: max_price
        type: integer
      - description: Поддержка почасовой аренды
        in: query
        name: rental_type_hourly
        type: boolean
      - description: Поддержка посуточной аренды
        in: query
        name: rental_type_daily
        type: boolean
      - description: Тип квартиры (ID)
        in: query
        name: apartment_type_id
        type: integer
      - description: Доступность квартиры (true - свободная, false - занятая)
        in: query
        name: is_free
        type: boolean
      - description: Статус квартиры
        in: query
        name: status
        type: string
      - default: 1
        description: Номер страницы
        in: query
        name: page
        type: integer
      - default: 10
        description: Размер страницы
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получение списка квартир
      tags:
      - apartments
    post:
      consumes:
      - application/json
      description: Создает новую квартиру для владельца
      parameters:
      - description: Данные квартиры
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.CreateApartmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Создание новой квартиры
      tags:
      - apartments
...
  /bookings:
    post:
      consumes:
      - application/json
      description: Создает новое бронирование квартиры
      parameters:
      - description: Данные бронирования
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CreateBookingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Создание бронирования
      tags:
      - bookings
...
  /chat/rooms:
    get:
      description: Получает список комнат чата пользователя с пагинацией
      parameters:
      - description: Фильтр по статусу комнаты (pending, active, closed)
        in: query
        name: status
        type: string
      - default: 1
        description: Номер страницы
        in: query
        name: page
        type: integer
      - default: 20
        description: Размер страницы
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список комнат чата
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/domain.ChatRoom'
                type: array
              pagination:
                type: object
              success:
                type: boolean
            type: object
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Список комнат чата
      tags:
      - chat
    post:
      consumes:
      - application/json
      description: Создает новую комнату чата для бронирования с консьержем
      parameters:
      - description: Данные для создания комнаты
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CreateChatRoomRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Комната чата создана
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.ChatRoom'
              type: object
        "400":
          description: Неверные данные запроса
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Создание комнаты чата
      tags:
      - chat
...
  /concierge/apartments:
    get:
      description: Получает список квартир, закрепленных за консьержем
      produces:
      - application/json
      responses:
        "200":
          description: Квартиры консьержа
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.Apartment'
                  type: array
              type: object
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получение квартир консьержа
      tags:
      - concierge-interface
  /concierge/bookings:
    get:
      description: Получает список бронирований для квартир консьержа
      parameters:
      - description: Фильтр по статусу
        in: query
        name: status
        type: string
      - description: Только активные брони
        in: query
        name: active
        type: boolean
      - default: 1
        description: Номер страницы
        in: query
        name: page
        type: integer
      - default: 20
        description: Размер страницы
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Брони консьержа
          schema:
            properties:
              data:
                properties:
                  bookings:
                    items:
                      $ref: '#/definitions/domain.Booking'
                    type: array
                  total:
                    type: integer
                type: object
              success:
                type: boolean
            type: object
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получение броней консьержа
      tags:
      - concierge-interface
  /concierge/chat/rooms:
    get:
      description: Получает список чат-комнат для консьержа
      parameters:
      - description: Фильтр по статусу
        in: query
        name: status
        type: string
      - default: 1
        description: Номер страницы
        in: query
        name: page
        type: integer
      - default: 20
        description: Размер страницы
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Чаты консьержа
          schema:
            properties:
              data:
                properties:
                  rooms:
                    items:
                      $ref: '#/definitions/domain.ChatRoom'
                    type: array
                  total:
                    type: integer
                type: object
              success:
                type: boolean
            type: object
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получение чатов консьержа
      tags:
      - concierge-interface
  /concierge/profile:
    get:
      description: Получает профиль текущего консьержа
      produces:
      - application/json
      responses:
        "200":
          description: Профиль консьержа
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.Concierge'
              type: object
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Консьерж не найден
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получение профиля консьержа
      tags:
      - concierge-interface
  /concierge/schedule:
    put:
      consumes:
      - application/json
      description: Обновляет рабочее расписание консьержа
      parameters:
      - description: Новое расписание
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.ConciergeSchedule'
      produces:
      - application/json
      responses:
        "200":
          description: Расписание обновлено
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Неверные данные запроса
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Обновление расписания консьержа
      tags:
      - concierge-interface
  /concierge/stats:
    get:
      description: Получает статистику работы консьержа
      produces:
      - application/json
      responses:
        "200":
          description: Статистика консьержа
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получение статистики консьержа
      tags:
      - concierge-interface
...
securityDefinitions:
  ApiKeyAuth:
    description: Bearer token для авторизации
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"